Enable fraud alerts and notifications
Enable fraud alerts and notifications for Microsoft Authenticator. "MFA fatigue attacks" have seen an increase in popularity recently. Some high profile breaches have been the result of them. The default configuration of an Azure AD tenant is such that an attacker who has phished a user's credentials can make an unlimited number of MFA push requests until the victim simply gives in. Even stopping short of an MFA fatigue attack, it may not occur to a user that an unsolicited prompt is indicative of their password being compromised.